backup/backend/requirements.txt
# =============================================================================
# TO_EXTRACTOR v6.2 - BACKEND REQUIREMENTS
# =============================================================================
# Aggiornato: 2025-01-03
# =============================================================================

# ===== Database PostgreSQL =====
psycopg2-binary==2.9.9      # Driver PostgreSQL (binario precompilato)

# ===== Framework Web =====
fastapi==0.109.0            # Framework API REST
uvicorn[standard]==0.27.0   # Server ASGI per FastAPI
python-multipart==0.0.6     # Upload file (UploadFile)

# ===== Validazione e Modelli =====
pydantic==2.5.3             # Validazione dati e modelli
pydantic-settings==2.1.0    # Gestione configurazioni
email-validator==2.2.0      # Validazione email (pydantic EmailStr)

# ===== Autenticazione e Sicurezza =====
PyJWT==2.8.0                # Token JWT
passlib[bcrypt]==1.7.4      # Hashing password (con backend bcrypt)

# ===== PDF Processing =====
pdfplumber==0.10.3          # Estrazione testo e tabelle da PDF

# ===== Gmail Monitor Integration =====
imapclient==2.3.1           # Connessione IMAP Gmail
pyzmail36==1.0.5            # Parsing email e allegati

# ===== Data Processing =====
pandas==2.1.4               # Import/export CSV e manipolazione dati
openpyxl==3.1.2             # Lettura/scrittura file Excel (.xlsx)

# ===== Fuzzy Matching =====
fuzzywuzzy==0.18.0          # Fuzzy string matching per lookup anagrafica
python-Levenshtein==0.23.0  # Acceleratore C per fuzzywuzzy

# ===== Utilities =====
python-dotenv==1.0.0        # Gestione file .env
python-json-logger==2.0.7   # Logging strutturato JSON
requests==2.31.0            # HTTP client per API calls
